requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deepinv==0.1.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastmri==0.3.0
filelock==3.9.0
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
h5py==3.10.0
hdf5storage==0.1.19
idna==3.4
imageio==2.34.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.28.1
runstats==2.0.0
scikit-image==0.22.0
scipy==1.12.0
sentry-sdk==1.40.3
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tifffile==2024.2.12
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchmetrics==1.3.1
torchvision==0.15.2+cu117
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.16.3
wcwidth==0.2.13
yarl==1.9.4